+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * if no matching properties are found for a property return an empty result set. See VertexiumNoMatchingPropertiesException
 * for SqlGraph, optimized SQL query used to retrieve vertices
+* fix softDeleteProperty on hidden elements
 
 # v2.2.2